Refactor Quiery Builders and Select Builders into their own files
