+++ source/Nevermore/Advanced/QueryBuilders/SourceQueryBuilder.cs
@@ -1,12 +1,11 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Nevermore.Querying;
 using Nevermore.Querying.AST;
 
-namespace Nevermore.Advanced
+namespace Nevermore.Advanced.QueryBuilders
 {
     // A union statement is built like a subquery.
     // This is because if you add other configuration (order by, where, column selection etc), then the inner query needs to be in a subquery
@@ -15,268 +14,6 @@ namespace Nevermore.Advanced
     //
     // This behaviour is different to normal subquery building,
     // because if the consumer explicitly asks for a subquery, it would be too presumptuous of us to not create a subquery for them under some circumstances
-    public class UnionSourceBuilder<TRecord> : SourceQueryBuilder<TRecord>, ISubquerySourceBuilder<TRecord> where TRecord : class
-    {
-        readonly ISelect select;
-        string alias;
-
-        public UnionSourceBuilder(ISelect select,
-            IReadQueryExecutor readQueryExecutor,
-            ITableAliasGenerator tableAliasGenerator,
-            IUniqueParameterNameGenerator uniqueParameterNameGenerator,
-            CommandParameterValues parameterValues,
-            Parameters parameters,
-            ParameterDefaults parameterDefaults)
-            : base(readQueryExecutor, tableAliasGenerator, uniqueParameterNameGenerator, parameterValues, parameters, parameterDefaults)
-        {
-            this.select = select;
-        }
-
-        protected override ISelectBuilder CreateSelectBuilder()
-        {
-            return new UnionSelectBuilder(select, alias, TableAliasGenerator);
-        }
-
-        public override IJoinSourceQueryBuilder<TRecord> Join(IAliasedSelectSource source, JoinType joinType, CommandParameterValues parameterValues, Parameters parameters, ParameterDefaults parameterDefaults)
-        {
-            return new JoinSourceQueryBuilder<TRecord>(AsSource(),
-                joinType,
-                source,
-                ReadQueryExecutor,
-                TableAliasGenerator,
-                UniqueParameterNameGenerator,
-                new CommandParameterValues(ParamValues, parameterValues),
-                new Parameters(Params, parameters),
-                new ParameterDefaults(ParamDefaults, parameterDefaults));
-        }
-
-        public ISubquerySource AsSource()
-        {
-            if (string.IsNullOrEmpty(alias))
-            {
-                Alias(TableAliasGenerator.GenerateTableAlias());
-            }
-            return new SubquerySource(select, alias);
-        }
-
-        public ISubquerySourceBuilder<TRecord> Alias(string subqueryAlias)
-        {
-            alias = subqueryAlias;
-            return this;
-        }
-    }
-
-    public class SubquerySourceBuilder<TRecord> : SourceQueryBuilder<TRecord>, ISubquerySourceBuilder<TRecord> where TRecord : class
-    {
-        readonly ISelect select;
-        string alias;
-
-        public SubquerySourceBuilder(ISelect select,
-            IReadQueryExecutor readQueryExecutor,
-            ITableAliasGenerator tableAliasGenerator,
-            IUniqueParameterNameGenerator uniqueParameterNameGenerator,
-            CommandParameterValues parameterValues,
-            Parameters parameters,
-            ParameterDefaults parameterDefaults)
-            : base(readQueryExecutor, tableAliasGenerator, uniqueParameterNameGenerator, parameterValues, parameters, parameterDefaults)
-        {
-            this.select = select;
-        }
-
-        protected override ISelectBuilder CreateSelectBuilder()
-        {
-            return new SubquerySelectBuilder(AsSource());
-        }
-
-        public override IJoinSourceQueryBuilder<TRecord> Join(IAliasedSelectSource source, JoinType joinType, CommandParameterValues parameterValues, Parameters parameters, ParameterDefaults parameterDefaults)
-        {
-            return new JoinSourceQueryBuilder<TRecord>(AsSource(),
-                joinType,
-                source,
-                ReadQueryExecutor,
-                TableAliasGenerator,
-                UniqueParameterNameGenerator,
-                new CommandParameterValues(ParamValues, parameterValues),
-                new Parameters(Params, parameters),
-                new ParameterDefaults(ParamDefaults, parameterDefaults));
-        }
-
-        public ISubquerySource AsSource()
-        {
-            if (string.IsNullOrEmpty(alias))
-            {
-                Alias(TableAliasGenerator.GenerateTableAlias());
-            }
-
-            return new SubquerySource(select, alias);
-        }
-
-        public ISubquerySourceBuilder<TRecord> Alias(string subqueryAlias)
-        {
-            alias = subqueryAlias;
-            return this;
-        }
-    }
-
-    public class JoinSourceQueryBuilder<TRecord> : SourceQueryBuilder<TRecord>, IJoinSourceQueryBuilder<TRecord> where TRecord : class
-    {
-        readonly IAliasedSelectSource originalSource;
-        readonly List<Join> intermediateJoins = new List<Join>();
-        JoinType type;
-        IAliasedSelectSource joinSource;
-        List<JoinClause> clauses;
-
-        public JoinSourceQueryBuilder(IAliasedSelectSource originalSource,
-            JoinType joinType,
-            IAliasedSelectSource nextJoin,
-            IReadQueryExecutor readQueryExecutor,
-            ITableAliasGenerator tableAliasGenerator,
-            IUniqueParameterNameGenerator uniqueParameterNameGenerator,
-            CommandParameterValues parameterValues,
-            Parameters parameters,
-            ParameterDefaults parameterDefaults)
-            : base(readQueryExecutor, tableAliasGenerator, uniqueParameterNameGenerator, parameterValues, parameters, parameterDefaults)
-        {
-            this.originalSource = originalSource;
-            clauses = new List<JoinClause>();
-            joinSource = nextJoin;
-            type = joinType;
-        }
-
-        protected override ISelectBuilder CreateSelectBuilder()
-        {
-            if (clauses.Count == 0)
-            {
-                throw new InvalidOperationException("Must have at least one 'ON' clause per join");
-            }
-            var joinedSource = new JoinedSource(originalSource, intermediateJoins.Concat(new [] {new Join(clauses.ToList(), joinSource, type)}).ToList());
-            return new JoinSelectBuilder(joinedSource);
-        }
-
-        public override IJoinSourceQueryBuilder<TRecord> Join(IAliasedSelectSource source, JoinType joinType, CommandParameterValues parameterValues, Parameters parameters, ParameterDefaults parameterDefaults)
-        {
-            if (clauses.Count == 0)
-            {
-                throw new InvalidOperationException("Must have at least one 'ON' clause per join");
-            }
-
-            intermediateJoins.Add(new Join(clauses.ToList(), joinSource, type));
-            clauses = new List<JoinClause>();
-            joinSource = source;
-            type = joinType;
-
-            var commandParameterValues = new CommandParameterValues(ParamValues, parameterValues);
-            var combinedParameters = new Parameters(Params, parameters);
-            var combinedParameterDefaults = new ParameterDefaults(ParamDefaults, parameterDefaults);
-
-            ParamValues.Clear();
-            ParamValues.AddRange(commandParameterValues);
-            Params.Clear();
-            Params.AddRange(combinedParameters);
-            ParamDefaults.Clear();
-            ParamDefaults.AddRange(combinedParameterDefaults);
-
-            return this;
-        }
-
-        public IJoinSourceQueryBuilder<TRecord> On(string leftField, JoinOperand operand, string rightField)
-        {
-            return On(originalSource.Alias, leftField, operand, rightField);
-        }
-
-        public IJoinSourceQueryBuilder<TRecord> On(string leftTableAlias, string leftField, JoinOperand operand, string rightField)
-        {
-            var newClause = new JoinClause(leftTableAlias, leftField, operand, joinSource.Alias, rightField);
-            clauses.Add(newClause);
-            return this;
-        }
-    }
-
-    public class TableSourceQueryBuilder<TRecord> : SourceQueryBuilder<TRecord>, ITableSourceQueryBuilder<TRecord> where TRecord : class
-    {
-        string tableOrViewName;
-        string alias;
-        string schemaName;
-        string idColumnName;
-
-        public TableSourceQueryBuilder(string tableOrViewName,
-            string schemaName,
-            string idColumnName,
-            IReadTransaction readQueryExecutor,
-            ITableAliasGenerator tableAliasGenerator,
-            IUniqueParameterNameGenerator uniqueParameterNameGenerator,
-            CommandParameterValues parameterValues,
-            Parameters parameters,
-            ParameterDefaults parameterDefaults)
-            : base(readQueryExecutor, tableAliasGenerator, uniqueParameterNameGenerator, parameterValues, parameters, parameterDefaults)
-        {
-            this.schemaName = schemaName;
-            this.tableOrViewName = tableOrViewName;
-            this.idColumnName = idColumnName;
-        }
-
-        protected override ISelectBuilder CreateSelectBuilder()
-        {
-            return new TableSelectBuilder(CreateSimpleTableSource(), new Column(idColumnName));
-        }
-
-        public override IJoinSourceQueryBuilder<TRecord> Join(IAliasedSelectSource source, JoinType joinType, CommandParameterValues parameterValues, Parameters parameters, ParameterDefaults parameterDefaults)
-        {
-            return new JoinSourceQueryBuilder<TRecord>(CreateAliasedTableSource(),
-                joinType,
-                source,
-                ReadQueryExecutor,
-                TableAliasGenerator,
-                UniqueParameterNameGenerator,
-                new CommandParameterValues(ParamValues, parameterValues),
-                new Parameters(Params, parameters),
-                new ParameterDefaults(ParamDefaults, parameterDefaults));
-        }
-
-        public ITableSourceQueryBuilder<TRecord> View(string viewName)
-        {
-            tableOrViewName = viewName;
-            return this;
-        }
-
-        public ITableSourceQueryBuilder<TRecord> Table(string tableName)
-        {
-            tableOrViewName = tableName;
-            return this;
-        }
-
-        public ITableSourceQueryBuilder<TRecord> Alias(string tableAlias)
-        {
-            alias = tableAlias;
-            return this;
-        }
-
-        public IAliasedSelectSource AsAliasedSource()
-        {
-            return CreateAliasedTableSource();
-        }
-
-        public IQueryBuilder<TRecord> Hint(string tableHint)
-        {
-            var source = new TableSourceWithHint(CreateSimpleTableSource(), tableHint);
-            return CreateQueryBuilder(new TableSelectBuilder(source, new Column(idColumnName)));
-        }
-
-        ISimpleTableSource CreateSimpleTableSource()
-        {
-            var columnNames = ReadQueryExecutor.GetColumnNames(schemaName, tableOrViewName);
-
-            return alias == null 
-                ? new SimpleTableSource(tableOrViewName, schemaName, columnNames) 
-                : new AliasedTableSource(new SimpleTableSource(tableOrViewName, schemaName, columnNames), alias);
-        }
-
-        AliasedTableSource CreateAliasedTableSource()
-        {
-            var columnNames = ReadQueryExecutor.GetColumnNames(schemaName, tableOrViewName);
-            return new AliasedTableSource(new SimpleTableSource(tableOrViewName, schemaName, columnNames), alias ?? TableAliasGenerator.GenerateTableAlias(tableOrViewName));
-        }
-    }
 
     public abstract class SourceQueryBuilder<TRecord> : IQueryBuilder<TRecord> where TRecord : class
     {
